refactor: replace ZeroPollProvider with EngineProvider
